Reoccurrence lock and some refactoring
timers now get parsed through a reoccurrence method 'Cull_DoomTimer_List' that tries to lock a name down for a good reading. seems to work good so far, have to add some more logic for timers that are ticking up from being inside the portal. and logic to enable the beeper and callouts