Boss HP Monitor Complete
The functionality for monitoring the boss hp and announces the mechanics + starts the doom timer monitor is fully implemented.
fixed a cpu sucking problem with the doom timers, with 4 doom timers monitoring the situation theyd suck up 20% of CPU resources. There was no thread sleep in the doom timer monitor thread. back down to around 10% cpu resources and sub 500mb of ram. although Ive notice the HP monitor thread has none or very little of any memory creep with the mats / bitmaps. possibly the doom thread could be optimized so the garbage collector does not need to be called as often or manually?
demo youtube video https://youtu.be/ZmA7TMTvO4Y